script

+++ admin/tools/ecs_pre.sh
@@ -2,6 +2,13 @@
 yum install git pigz docker ntp firewalld -y
 yum remove *nfs* -y
 systemctl disable rpcbind
-shutdown -r now
+echo "${1}"
+
+if  [ "${1}" == "reboot" ];then
+        shutdown -r now
+else
+    sh ./ecs.sh
+fi
+
